update zfs submodule to 2.2.7 and refresh patches
drop the cherry-picked kernel 6.11 compatibility patches.

this upstream release [0] is a bit larger than usual, and might have a bit
of potential for regressions.

apart from the compatibilty fixes for newer kernels (up until 6.12),
fixes to the testsystem,ci and fixes for FreeBSD the following might be
interesting for our use-cases:
* 308d04ac3 ("Fix inconsistent mount options for ZFS root")
* bc0d89bfc ("Fix an uninitialized data access (#16511)")
* f4e66db40 ("vdev_disk: move abd return and free off the interrupt handler")
* 0f86fcc2a ("Linux: Fix zfs_prune panics")
* cf80a803d ("zvol: ensure device minors are properly cleaned up")
* 299da6ace ("Fix race in libzfs_run_process_impl")
* 0bd8481aa ("add get_name implementation for exports. (#16833)")

I did some minimal testing on 2 VM's with a few nested containers and
storage replication, additionally I installed that version on a VM with
/ on ZFS - mostly for checking 308d04ac3 in our environment.

[0] https://github.com/openzfs/zfs/releases/tag/zfs-2.2.7

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>